+++ mocker-back/src/main/java/com/mocker/service/impl/TableServiceImpl.java
@@ -1,22 +1,21 @@
 package com.mocker.service.impl;
 
+import com.mocker.configuration.security.ApplicationContextHolder;
+import com.mocker.domain.exception.BadRequestException;
 import com.mocker.domain.exception.NotFoundException;
 import com.mocker.domain.model.entity.Field;
 import com.mocker.domain.model.entity.Option;
+import com.mocker.domain.model.entity.Schema;
 import com.mocker.domain.model.entity.Table;
-import com.mocker.repository.FieldRepository;
-import com.mocker.repository.OptionRepository;
-import com.mocker.repository.TableRelationRepository;
-import com.mocker.repository.TableRepository;
+import com.mocker.domain.model.entity.enumeration.Role;
+import com.mocker.repository.*;
+import com.mocker.service.PermissionService;
 import com.mocker.service.TableService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 
 /**
  * @author Luong Quoc Trung
@@ -30,21 +29,36 @@ public class TableServiceImpl implements TableService {
     private final FieldRepository fieldRepository;
     private final OptionRepository optionRepository;
     private final TableRelationRepository tableRelationRepository;
+    private final PermissionService permissionService;
 
     @Override
     public Table getTable(UUID id) {
-        return Optional.ofNullable(tableRepository.findOneFetchFields(id))
-                .orElseThrow(() -> new NotFoundException(id));
+        permissionService.checkPermission(id, Table.class);
+        return tableRepository.findOneFetchFields(id);
     }
 
     @Override
     public List<Table> getTablesBySchema(UUID schemaId) {
-        return Optional.ofNullable(tableRepository.findAllBySchemaFetchFields(schemaId))
-                .orElseThrow(() -> new NotFoundException(schemaId));
+        permissionService.checkPermission(schemaId, Schema.class);
+        return tableRepository.findAllBySchemaFetchFields(schemaId);
     }
 
     @Override
     public Table upsert(Table table) {
+        permissionService.checkPermission(
+                table.getSchema().getId(),
+                Schema.class,
+                List.of(Role.GROUP_ADMIN, Role.GROUP_ASSOCIATE),
+                "You can not be allowed to perform this action!<br/>Please try again later when you have a new role with <b>group admin</b> or <b>group associate</b>.");
+        Table existingTableName = tableRepository.findOneByNameAndSchemaId(table.getName(), table.getSchema().getId());
+        if (Objects.nonNull(existingTableName)) {
+            if (Objects.isNull(table.getId())) {
+                // Create a new table
+                throw new BadRequestException("The table with name " + table.getName() + " is already existed, please create with the different table name");
+            } else if (!Objects.equals(existingTableName.getId(), table.getId())) {
+                throw new BadRequestException("The table with name " + table.getName() + "is already existed, please update it with the different table name");
+            }
+        }
         if (table.getId() != null) {
             List<UUID> ids = table.getFields().stream().map(Field::getId).toList();
             List<UUID> fieldIdsNeedToRemove = tableRepository.findOneFetchFields(table.getId())
@@ -76,6 +90,11 @@ public class TableServiceImpl implements TableService {
 
     @Override
     public Table delete(UUID id) {
+        permissionService.checkPermission(
+                id,
+                Table.class,
+                List.of(Role.GROUP_ADMIN, Role.GROUP_ASSOCIATE),
+                "You can not be allowed to perform this action!<br/>Please try again later when you have a new role with <b>group admin</b> or <b>group associate</b>.");
         Table table = Optional.ofNullable(tableRepository.findOneFetchFields(id)).orElseThrow(() -> new NotFoundException(id));
         optionRepository.deleteAll(table.getFields()
                 .stream()
